flute rename lib and dir from sst to stt
