+++ src/CMakeLists.txt
@@ -118,7 +118,7 @@ add_subdirectory(OpenDB)
 add_subdirectory(sta)
 add_subdirectory(dbSta)
 add_subdirectory(rsz)
-add_subdirectory(sst)
+add_subdirectory(stt)
 add_subdirectory(replace)
 add_subdirectory(dpl)
 add_subdirectory(fin)
@@ -167,7 +167,7 @@ target_link_libraries(openroad
   fin
   rsz
   ppl
-  sst
+  stt
   dbSta
   OpenSTA
   opendb
