+++ src/TritonRoute/CMakeLists.txt
@@ -134,7 +134,7 @@ target_link_libraries(TritonRoute
   PUBLIC
     gui
     opendb
-    sst
+    stt
     OpenSTA
     utl
     Threads::Threads
