+++ src/stt/LICENSE
