+++ src/stt/etc/MakeDatVar.tcl
